moved task into init
